% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasa_label_db.R
\name{.lasa_empty_label_db}
\alias{.lasa_empty_label_db}
\title{An empty, correctly-shaped \code{lasa_label_db} object}
\usage{
.lasa_empty_label_db()
}
\description{
An empty, correctly-shaped \code{lasa_label_db} object
}
\keyword{internal}
