% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasa_varinfo_parse.R
\name{.lasa_varinfo_pdf_tokens}
\alias{.lasa_varinfo_pdf_tokens}
\title{Read one PDF's tokens into a single data frame}
\usage{
.lasa_varinfo_pdf_tokens(path)
}
\description{
Read one PDF's tokens into a single data frame
}
\keyword{internal}
